+++ src/firenvim
@@ -17,11 +17,13 @@ password = random.getrandbits(64)
 firenvim = subprocess.Popen(["/home/me/.local/share/firenvim/firenvim"], bufsize = 0, stdout = subprocess.PIPE, stdin = subprocess.PIPE, stderr = subprocess.PIPE)
 request = 'abcde{"newInstance":true,"password":"' + str(password) + '"}'
 firenvim.stdin.write(bytes(request, 'utf-8'))
-outs = firenvim.stdout.read(40)
-result = json.loads(outs[4:])
+out = b"";
+while not out.endswith(b"}"):
+    out += firenvim.stdout.read(1)
+result = json.loads(out[4:])
 
-with open(dir("firenvim.js"), "rb") as f:
-    js = f.read().replace(
+with open(dir("frame.js"), "rb") as f:
+    framejs = f.read().replace(
         b'Promise.resolve({ port: "QUTEBROWSER_PORT", password: "QUTEBROWSER_PASSWORD" });',
         bytes('Promise.resolve({ port: %s, password: "%s" });' % (result["port"], password), 'utf-8'),
     )
@@ -30,17 +32,19 @@ with open(dir("index.html"), "rb") as f:
     html = f.read()
     script_tag = html.find(b'<script type="application/javascript" src="index.js">')
     script_tag_end = html[script_tag:].find(b'</script>')
-    html = html[0:script_tag] + b'<script>\n' + js + b'\n</script>' + html[script_tag_end:]
+    html = html[0:script_tag] + b'<script>\n' + framejs + b'\n</script>' + html[script_tag_end:]
     with open("/tmp/blah", "wb") as f2:
         f2.write(html)
 
+with open(dir("content.js"), "rb") as f:
+    uint8_array = bytes(''.join([str(c) + "," for c in html]), 'utf-8')
+    contentjs = f.read().replace(
+        b"/* QUTEBROWSER_PAGE */",
+        b"(new Uint8Array([" + uint8_array + b"])).buffer"
+    )
+
 with open(dir("index.js"), "wb") as f:
-    f.write(b'const iframe = document.createElement("iframe");\n')
-    f.write(b'const blob = new Blob([new Uint8Array([')
-    f.write(bytes(''.join([str(c) + "," for c in html]), 'utf-8'))
-    f.write(b']).buffer], {type: "text/html; charset=utf-8"});\n')
-    f.write(b'iframe.src = URL.createObjectURL(blob);\n')
-    f.write(b'document.documentElement.appendChild(iframe);\n')
+    f.write(contentjs)
 
 with open(os.environ["QUTE_FIFO"], "w") as f:
     f.write("jseval --file %s" % dir("index.js"))
